refactoring autoDrive tool
