+++ src/agents/tools/autonomysAutoDrive/utils/utils.ts
@@ -1,0 +1,3 @@
+import { createLogger } from '../../../../utils/logger.js';
+
+export const logger = createLogger('auto-drive-tools');
